x/od/types/message_propose_send_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"opendao/testutil/sample"
)

func TestMsgProposeSend_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgProposeSend
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgProposeSend{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgProposeSend{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
